+++ src/Test/intlinear.jl
@@ -53,6 +53,9 @@ function int1test(model::MOI.ModelLike, config::TestConfig)
     @test MOI.get(model, MOI.ObjectiveSense()) == MOI.MaxSense
 
     if config.solve
+        @test MOI.get(model, MOI.TerminationStatus()) == MOI.OptimizeNotCalled
+        @test MOI.get(model, MOI.PrimalStatus()) == MOI.NoSolution
+
         MOI.optimize!(model)
 
         @test MOI.get(model, MOI.TerminationStatus()) == MOI.Success
@@ -121,6 +124,9 @@ function int2test(model::MOI.ModelLike, config::TestConfig)
         @test MOI.get(model, MOI.ObjectiveSense()) == MOI.MaxSense
 
         if config.solve
+            @test MOI.get(model, MOI.TerminationStatus()) == MOI.OptimizeNotCalled
+            @test MOI.get(model, MOI.PrimalStatus()) == MOI.NoSolution
+
             MOI.optimize!(model)
 
             @test MOI.get(model, MOI.TerminationStatus()) == MOI.Success
@@ -206,6 +212,9 @@ function int2test(model::MOI.ModelLike, config::TestConfig)
         @test MOI.get(model, MOI.ObjectiveSense()) == MOI.MaxSense
 
         if config.solve
+            @test MOI.get(model, MOI.TerminationStatus()) == MOI.OptimizeNotCalled
+            @test MOI.get(model, MOI.PrimalStatus()) == MOI.NoSolution
+
             MOI.optimize!(model)
 
             @test MOI.get(model, MOI.TerminationStatus()) == MOI.Success
@@ -275,6 +284,9 @@ function int3test(model::MOI.ModelLike, config::TestConfig)
     MOI.set(model, MOI.ObjectiveSense(), MOI.MaxSense)
 
     if config.solve
+        @test MOI.get(model, MOI.TerminationStatus()) == MOI.OptimizeNotCalled
+        @test MOI.get(model, MOI.PrimalStatus()) == MOI.NoSolution
+
         MOI.optimize!(model)
 
         @test MOI.get(model, MOI.TerminationStatus()) == MOI.Success
@@ -331,6 +343,9 @@ function knapsacktest(model::MOI.ModelLike, config::TestConfig)
     end
 
     if config.solve
+        @test MOI.get(model, MOI.TerminationStatus()) == MOI.OptimizeNotCalled
+        @test MOI.get(model, MOI.PrimalStatus()) == MOI.NoSolution
+
         MOI.optimize!(model)
 
         @test MOI.get(model, MOI.TerminationStatus()) == MOI.Success